refactor(core): some runtime methods should live on the module map
